# Populate warning data for batch insert warnings with JSON source + populate line number for JSON scan errors