--- a/extension/json/src/include/json_extension.h
+++ b/extension/json/src/include/json_extension.h
@@ -8,6 +8,7 @@ namespace json_extension {
 class JsonExtension final : public extension::Extension {
 public:
     static constexpr char JSON_TYPE_NAME[] = "json";
+    static constexpr common::idx_t JSON_SCAN_FILE_IDX = 0;
 
     static void load(main::ClientContext* context);
 };