--- a/extension/json/src/include/json_scan.h
+++ b/extension/json/src/include/json_scan.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "function/table/bind_data.h"
 #include "function/table_functions.h"
 
 namespace kuzu {