--- a/src/include/common/data_chunk/data_chunk.h
+++ b/src/include/common/data_chunk/data_chunk.h
@@ -33,7 +33,7 @@ public:
 
     inline uint32_t getNumValueVectors() const { return valueVectors.size(); }
 
-    inline std::shared_ptr<ValueVector> getValueVector(uint64_t valueVectorPos) {
+    inline const std::shared_ptr<ValueVector>& getValueVector(uint64_t valueVectorPos) const {
         return valueVectors[valueVectorPos];
     }