… dataset/copy-fault-tests/invalid-row/vPersonLongRow.csv
… dataset/json-error/index_lookup_error.json
… dataset/json-error/newline_delimited_invalid_value.json
… dataset/json-error/person_ok.json
… dataset/json-error/structured_trailing_comma.json
… dataset/json-error/unstructured.json
… extension/CMakeLists.txt
… extension/json/test/error.test
… extension/json/test/scan_json.test
… scripts/headers.txt
… src/include/common/type_utils.h
… src/include/common/utils.h
… src/include/function/table/bind_data.h
… src/include/function/table_functions.h
… src/include/processor/execution_context.h
… src/include/processor/operator/persistent/index_builder.h
… src/include/processor/operator/persistent/reader/copy_from_error.h
… src/include/processor/operator/persistent/reader/csv/base_csv_reader.h
… src/include/processor/operator/persistent/reader/csv/driver.h
… src/include/processor/operator/persistent/reader/csv/parallel_csv_reader.h
… src/include/processor/operator/persistent/reader/csv/serial_csv_reader.h
… src/include/processor/operator/persistent/reader/file_error_handler.h
… src/include/processor/operator/persistent/reader/rdf/rdf_scan.h
… src/include/processor/warning_context.h
… src/processor/operator/index_lookup.cpp
… src/processor/operator/persistent/batch_insert_error_handler.cpp
… src/processor/operator/persistent/index_builder.cpp
… src/processor/operator/persistent/reader/CMakeLists.txt
… src/processor/operator/persistent/reader/copy_from_error.cpp
… src/processor/operator/persistent/reader/csv/CMakeLists.txt