# Description
Batch insert warnings (e.g. duplicate PK, index lookup failure) from JSON source now have warning info (e.g. line number, file path) populated in the warning table.

As part of this change, the column `skipped_line` in the warning table have been renamed to `skipped_line_or_record` as for JSON we report the record that causes the error (which can span multiple lines, or only part of a line). The new warning table will look something like (if parsed from a JSON file):
```
┌──────────┬───────────────────────────────────────────────────────────────────────────────────────────────────────────┬─────────────────────────────────────────┬─────────────┬────────────────────────┐
│ query_id │ message                                                                                                   │ file_path                               │ line_number │ skipped_line_or_record │
│ UINT64   │ STRING                                                                                                    │ STRING                                  │ UINT64      │ STRING                 │
├──────────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────┼─────────────────────────────────────────┼─────────────┼────────────────────────┤
│ 2        │ Found duplicated primary key value 1, which violates the uniqueness constraint of the primary key column. │ /home/kuzu/test/csv/dataset-json.json   │ 9200001     │ {"ID": 1, "age": 2}    │
│ 2        │ Found duplicated primary key value 1, which violates the uniqueness constraint of the primary key column. │ /home/kuzu/test/csv/dataset-json.json   │ 9000001     │ {"ID": 1, "age": 2}    │
│ 2        │ Found duplicated primary key value 1, which violates the uniqueness constraint of the primary key column. │ /home/kuzu/test/csv/dataset-json.json   │ 9100001     │ {"ID": 1, "age": 2}    │
└──────────┴───────────────────────────────────────────────────────────────────────────────────────────────────────────┴─────────────────────────────────────────┴─────────────┴────────────────────────┘
```
JSON parsing messages are also updated to include line number + the reconstructed skipped line, something like `Copy exception: Error in file ${KUZU_ROOT_DIRECTORY}/dataset/json-error/structured.json on line 3: Malformed JSON: unexpected character, expected a valid JSON value. Line/record containing the error: '{"c":}...'`.

Most of the changes in this PR are refactorings to support the above changes.

**Sanity benchmarks are [here](https://github.com/kuzudb/kuzu/pull/4294#issuecomment-2374377609)**

Fixes #4262

# Contributor agreement

- [ ] I have read and agree to the [Contributor Agreement](https://github.com/kuzudb/kuzu/blob/master/CLA.md).