@royi-luo 2024-09-25
Some quick benchmarks, since there is quite a bit of refactoring I tested on multiple datasets to check for performance regressions. Note that due to some additional optimizations for CSV parsing results may be better than on master, although there is some additional copying overhead from `WarningSourceData` storing type information for each value:

## Dataset
Schema:
```
CREATE NODE TABLE person(ID UINT32, PRIMARY KEY(ID));
CREATE REL TABLE knows(FROM person TO person, MANY_MANY);
```
Details: 10M nodes in `person`, 10M rels in `knows`

## Duplicate PK Errors (CSV)

Query: `COPY person FROM "/home/kuzu/test/dataset/many-ok-pk.csv"(IGNORE_ERRORS={value});`

### Results
Branch|`IGNORE_ERRORS`|Error %|Copy Time (master)|Copy Time (new)
----|----|----|----|----
new|false|0%|515ms|575ms
new|true|0%|708ms|742ms
new|true|1%|852ms|893ms
new|true|10%|1072ms|1210ms
master|true|100%|4701ms|4944ms

## Index Lookup Errors (CSV)

Copy query: `COPY knows FROM "/home/kuzu/test/dataset/missing-pk-rels.csv"(IGNORE_ERRORS={value});`

### Results
Branch|`IGNORE_ERRORS`|Error %|Copy Time (master)|Copy Time (new)
----|----|----|----|----
new|false|0%|1546ms|1629ms
new|true|0%|2329ms|2352ms
new|true|10%|2256ms|2184ms
new|true|100%|1381ms|1268ms

## Duplicate PK Errors (JSON)